requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
constantly==15.1.0
cryptography==35.0.0
cycler==0.10.0
dateparser==1.1.0
greenlet==1.1.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
multidict==5.2.0
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.8
requests==2.25.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.25
Twisted==21.2.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.0
yfinance==0.1.63
zope.interface==5.4.0
